displays: drop the dead PnP-name scrapers
Html.cs carried three scrapers that looked a monitor's PnP code up on
driverlookup.com, driveragent.com and driversdownloader.com, plus the
HttpWebRequest plumbing behind them — spoofed user agent, cookie container,
commented-out __VIEWSTATE handling. An old attempt at enriching monitor names
that has not worked in years.

Nothing called any of it. The single live caller
(WindowsLayoutFactory.SetPnpDeviceName) uses CleanupPnpName, a local string
tidy-up that never touches the network, so this deletes the scrapers and keeps
the cleanup.

Worth removing rather than leaving inert: a grep of the tree for outbound URLs
turned these up while writing the privacy policy, and code aimed at scraping
third-party sites reads badly in an application that otherwise only contacts
GitHub and televisions on the local network — whatever the call graph says.

The remaining function is renamed with its file, HtmlHelper.CleanupPnpName ->
PnpName.Cleanup: nothing here parses HTML any more.
